fix: popup refactoring
