+++ src/components/NotFound/NotFound.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { ROUTES_MAP } from '../../utils/routesMap';
 import './NotFound.css';
 
-function NotFound({ login }) {
+function NotFound() {
   return (
     <div className='notFound'>
       <h1 className='notFound__title'>404</h1>
