+++ src/components/Popups/PopupWithForm/PopupWithForm.css
@@ -66,7 +66,7 @@
   padding: 0;
   top: -40px;
   right: -40px;
-  background-image: url(../../images/popup/back.svg);
+  background-image: url(../../../images/popup/back.svg);
   background-color: transparent;
   background-repeat: no-repeat;
   background-position: center;
@@ -92,34 +92,4 @@
   opacity: 1;
   transition: visibility 0s, opacity 0.5s linear;
 }
-.popup__button {
-  font-family: Roboto;
-  font-weight: 500;
-  font-style: normal;
-  font-size: 18px;
-  text-align: center;
-  width: 360px;
-  height: 64px;
-  color: #ffffff;
-  background: #2f71e5;
-  box-shadow: 0px 5px 15px rgba(14, 26, 57, 0.2);
-  border-radius: 100px;
-  border: none;
-  cursor: pointer;
-  outline: none;
-}
-.popup__button:hover {
-  background: #347eff;
-}
-.popup__button:active {
-  background: #2a65cc;
-}
-.popup__button_disable {
-  cursor: default;
-  background: #e6e8eb;
-}
-.popup__form {
-  display: flex;
-  flex-direction: column;
-  margin: 0 auto;
-}
+
